+++ client-spark/spark3/src/main/java/org/apache/spark/shuffle/writer/RssShuffleWriter.java
@@ -199,7 +199,7 @@ public class RssShuffleWriter<K, V, C> extends ShuffleWriter<K, V> {
       String appId,
       int shuffleId,
       String taskId,
-      long taskAttemptId,
+      int taskAttemptId,
       ShuffleWriteMetrics shuffleWriteMetrics,
       RssShuffleManager shuffleManager,
       SparkConf sparkConf,
