+++ storage/src/test/java/org/apache/uniffle/storage/HadoopShuffleHandlerTestBase.java
@@ -49,7 +49,7 @@ public class HadoopShuffleHandlerTestBase {
       HadoopShuffleWriteHandler writeHandler,
       int num,
       int length,
-      long taskAttemptId,
+      int taskAttemptId,
       Map<Long, byte[]> expectedData)
       throws Exception {
     List<ShufflePartitionedBlock> blocks = Lists.newArrayList();
@@ -71,7 +71,7 @@ public class HadoopShuffleHandlerTestBase {
       int partitionId,
       int num,
       int length,
-      long taskAttemptId,
+      int taskAttemptId,
       Map<Long, byte[]> expectedData,
       Map<Integer, List<ShufflePartitionedBlock>> expectedBlocks,
       Map<Integer, List<FileBasedShuffleSegment>> expectedIndexSegments,
